toy_grid_world/requirements.txt
numpy~=1.22.3
termcolor~=1.1.0
pytorch~=1.7.0
gym~=0.21.0
tensorboard~=2.6.0
pandas~=1.4.2
tqdm~=4.64.0
numba~=0.56.4